+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 	- Unstable Nuggets require an inverted Sigil
 - Which Angel Ring wing you have is now a Data Attachment
 	- It was previously a packet that modified a static `Map<UUID, Type>` which is gross
+- Items with the `#excessive_utilities:not_yet_implemented` now say it in the tooltip that they're NYI
 
 ### Fixed